+++ Sources/App/Helper/FlatBufferWriter/FlatBuffers+ForecastApi.swift
@@ -1,131 +1,6 @@
 import Foundation
 import FlatBuffers
-import Vapor
 
-extension ForecastapiResult {
-    /// Convert data into a FlatBuffers scheme far fast binary encoding and transfer
-    /// Each `ForecastapiResult` is converted indifuavually into an flatbuffer message -> very long time-series require a lot of memory
-    /// Data is using `size prefixed` flatbuffers to allow streaming of multiple messages for multiple locations
-    func toFlatbuffersResponse(fixedGenerationTime: Double?) throws -> Response {
-        // First excution outside stream, to capture potential errors better
-        //var first = try self.first?()
-        if results.count > 1000 {
-            throw ForecastapiError.generic(message: "Only up to 1000 locations can be requested at once")
-        }
-        let response = Response(body: .init(stream: { writer in
-            writer.submit {
-                // TODO: Zero-copy for flatbuffer to NIO bytebuffer conversion. Probably writing an optimised flatbuffer encoder would be better.
-                // TODO: Estimate initial buffer size
-                let initialSize = Int32(4096) // Int32(((first?.estimatedFlatbufferSize ?? 4096)/4096+1)*4096)
-                var fbb = FlatBufferBuilder(initialSize: initialSize)
-                var b = BufferAndWriter(writer: writer)
-                //if let first {
-                //    first.writeToFlatbuffer(&fbb)
-                //    b.buffer.writeBytes(fbb.buffer.unsafeRawBufferPointer)
-                //    fbb.clear()
-                //}
-                //first = nil
-                try await b.flushIfRequired()
-                for location in results {
-                    for model in location.results {
-                        try Model.writeToFlatbuffer(section: model, &fbb, timezone: location.timezone, fixedGenerationTime: fixedGenerationTime)
-                        b.buffer.writeBytes(fbb.buffer.unsafeRawBufferPointer)
-                        fbb.clear()
-                        try await b.flushIfRequired()
-                    }
-                }
-                try await b.flush()
-                try await b.end()
-            }
-        }))
-        response.headers.replaceOrAdd(name: .contentType, value: "application/octet-stream")
-        return response
-    }
-}
-
-fileprivate extension FlatBuffers.ByteBuffer {
-    /// Create a pointer to the data region. Flatbuffer is filling the buffer backwards.
-    var unsafeRawBufferPointer: UnsafeRawBufferPointer {
-        .init(start: memory.advanced(by: self.capacity - Int(self.size)), count: Int(size))
-    }
-}
-
-extension ForecastapiResult {
-    static func encode(section: ApiSection<Model.DailyVariable>, _ fbb: inout FlatBufferBuilder) -> [Offset] {
-        let offsets: [Offset] = section.columns.map { v in
-            switch v.variables[0] {
-            case .float(let float):
-                return com_openmeteo_ValuesAndUnit.createValuesAndUnit(&fbb, valuesVectorOffset: fbb.createVector(float), unit: v.unit)
-            case .timestamp(let time):
-                return fbb.createVector(time.map({$0.timeIntervalSince1970}))
-            }
-        }
-        return offsets
-    }
-    
-    
-    static func encode(section: ApiSection<SurfaceAndPressureVariable>, _ fbb: inout FlatBufferBuilder) -> (surface: [(variable: Model.HourlyVariable, offset: Offset)], pressure: [(variable: Model.HourlyPressureType, offset: Offset)]) {
-        var surfaces = [(variable: Model.HourlyVariable, offset: Offset)]()
-        surfaces.reserveCapacity(section.columns.count)
-        var pressures = [(variable: Model.HourlyPressureType, unit: SiUnit, offsets: [Offset])]()
-        
-        for v in section.columns {
-            switch v.variable {
-            case .surface(let surface):
-                let offset = com_openmeteo_ValuesAndUnit.createValuesAndUnit(&fbb, valuesVectorOffset: v.variables[0].expectFloatArray(&fbb), unit: v.unit)
-                surfaces.append((surface, offset))
-            case .pressure(let pressure):
-                let offset = com_openmeteo_ValuesAndLevel.createValuesAndLevel(&fbb, level: Int32(pressure.level), valuesVectorOffset: v.variables[0].expectFloatArray(&fbb))
-                if let pos = pressures.firstIndex(where: {$0.variable == pressure.variable}) {
-                    pressures[pos].offsets.append(offset)
-                } else {
-                    pressures.append((pressure.variable, v.unit, [offset]))
-                }
-            }
-        }
-        
-        let pressureVectors: [(variable: Model.HourlyPressureType, offset: Offset)] = pressures.map { (variable, unit, offsets) in
-            return (variable, com_openmeteo_ValuesUnitPressureLevel.createValuesUnitPressureLevel(&fbb, unit: unit, valuesVectorOffset: fbb.createVector(ofOffsets: offsets)))
-        }
-        
-        return (surfaces, pressureVectors)
-    }
-    
-    
-    /// Encode hourly variable for surface and pressure variables
-    static func encodeEnsemble(section: ApiSection<SurfaceAndPressureVariable>, _ fbb: inout FlatBufferBuilder) -> (surface: [(variable: Model.HourlyVariable, offset: Offset)], pressure: [(variable: Model.HourlyPressureType, offset: Offset)]) {
-        var surfaces = [(variable: Model.HourlyVariable, offset: Offset)]()
-        surfaces.reserveCapacity(section.columns.count)
-        var pressures = [(variable: Model.HourlyPressureType, unit: SiUnit, offsets: [Offset])]()
-        
-        for v in section.columns {
-            switch v.variable {
-            case .surface(let surface):
-                let oo = v.variables.enumerated().map { (member, data) in
-                    return com_openmeteo_ValuesAndMember.createValuesAndMember(&fbb, member: Int32(member), valuesVectorOffset: data.expectFloatArray(&fbb))
-                }
-                let offset = com_openmeteo_ValuesUnitAndMember.createValuesUnitAndMember(&fbb, unit: v.unit, valuesVectorOffset: fbb.createVector(ofOffsets: oo))
-                surfaces.append((surface, offset))
-            case .pressure(let pressure):
-                let oo = v.variables.enumerated().map { (member, data) in
-                    return com_openmeteo_ValuesAndMember.createValuesAndMember(&fbb, member: Int32(member), valuesVectorOffset: data.expectFloatArray(&fbb))
-                }
-                let offset = com_openmeteo_ValuesAndLevelAndMember.createValuesAndLevelAndMember(&fbb, level: Int32(pressure.level), valuesVectorOffset: fbb.createVector(ofOffsets: oo))
-                if let pos = pressures.firstIndex(where: {$0.variable == pressure.variable}) {
-                    pressures[pos].offsets.append(offset)
-                } else {
-                    pressures.append((pressure.variable, v.unit, [offset]))
-                }
-            }
-        }
-        
-        let pressureVectors: [(variable: Model.HourlyPressureType, offset: Offset)] = pressures.map { (variable, unit, offsets) in
-            return (variable, com_openmeteo_ValuesUnitPressureLevelAndMember.createValuesUnitPressureLevelAndMember(&fbb, unit: unit, valuesVectorOffset: fbb.createVector(ofOffsets: offsets)))
-        }
-        
-        return (surfaces, pressureVectors)
-    }
-}
 
 extension MultiDomains: ModelFlatbufferSerialisable {
     typealias HourlyVariable = ForecastSurfaceVariable
@@ -666,204 +541,5 @@ extension MultiDomains: ModelFlatbufferSerialisable {
         case .arome_france_hd:
             return .meteofranceAromeFranceHd
         }
-    }
-}
-
-extension ApiArray {
-    func expectFloatArray(_ fbb: inout FlatBuffers.FlatBufferBuilder) -> Offset {
-        switch self {
-        case .float(let array):
-            return fbb.createVector(array)
-        case .timestamp(_):
-            fatalError("Expected float array and not timestamps")
-        }
-    }
-}
-
-
-
-extension ApiSection {
-    func timeFlatBuffers() -> com_openmeteo_TimeRange {
-        return .init(
-            start: Int64(time.range.lowerBound.timeIntervalSince1970),
-            end: Int64(time.range.upperBound.timeIntervalSince1970),
-            interval: Int32(Int64(time.dtSeconds))
-        )
-    }
-}
-
-
-
-extension EnsembleMultiDomains: ModelFlatbufferSerialisable {
-    typealias HourlyVariable = EnsembleSurfaceVariable
-    
-    typealias HourlyPressureType = EnsemblePressureVariableType
-    
-    typealias DailyVariable = ForecastVariableDaily
-    
-    var flatBufferModel: com_openmeteo_EnsembleModel {
-        switch self {
-        case .icon_seamless:
-            return .iconSeamless
-        case .icon_global:
-            return .iconGlobal
-        case .icon_eu:
-            return .iconEu
-        case .icon_d2:
-            return .iconD2
-        case .ecmwf_ifs04:
-            return .ecmwfIfs04
-        case .gem_global:
-            return .gemGlobal
-        case .gfs_seamless:
-            return .gfsSeamless
-        case .gfs025:
-            return .gfs025
-        case .gfs05:
-            return .gfs025
-        }
-    }
-    
-    static func encodeHourly(section: ApiSection<ForecastapiResult<Self>.SurfaceAndPressureVariable>, _ fbb: inout FlatBufferBuilder) -> Offset {
-        let offsets = ForecastapiResult.encode(section: section, &fbb)
-        let start = com_openmeteo_EnsembleHourly.startEnsembleHourly(&fbb)
-        com_openmeteo_EnsembleHourly.add(time: section.timeFlatBuffers(), &fbb)
-        for (surface, offset) in offsets.surface {
-            switch surface {
-            case .weathercode:
-                com_openmeteo_EnsembleHourly.add(weathercode: offset, &fbb)
-            case .temperature_2m:
-                com_openmeteo_EnsembleHourly.add(temperature2m: offset, &fbb)
-            case .temperature_80m:
-                com_openmeteo_EnsembleHourly.add(temperature80m: offset, &fbb)
-            case .temperature_120m:
-                com_openmeteo_EnsembleHourly.add(temperature120m: offset, &fbb)
-            case .cloudcover:
-                com_openmeteo_EnsembleHourly.add(cloudcover: offset, &fbb)
-            case .pressure_msl:
-                com_openmeteo_EnsembleHourly.add(pressureMsl: offset, &fbb)
-            case .relativehumidity_2m:
-                com_openmeteo_EnsembleHourly.add(relativehumidity2m: offset, &fbb)
-            case .precipitation:
-                com_openmeteo_EnsembleHourly.add(precipitation: offset, &fbb)
-            case .rain:
-                com_openmeteo_EnsembleHourly.add(rain: offset, &fbb)
-            case .windgusts_10m:
-                com_openmeteo_EnsembleHourly.add(windgusts10m: offset, &fbb)
-            case .dewpoint_2m:
-                com_openmeteo_EnsembleHourly.add(dewpoint2m: offset, &fbb)
-            case .diffuse_radiation:
-                com_openmeteo_EnsembleHourly.add(diffuseRadiation: offset, &fbb)
-            case .direct_radiation:
-                com_openmeteo_EnsembleHourly.add(directRadiation: offset, &fbb)
-            case .apparent_temperature:
-                com_openmeteo_EnsembleHourly.add(apparentTemperature: offset, &fbb)
-            case .windspeed_10m:
-                com_openmeteo_EnsembleHourly.add(windspeed10m: offset, &fbb)
-            case .winddirection_10m:
-                com_openmeteo_EnsembleHourly.add(winddirection10m: offset, &fbb)
-            case .windspeed_80m:
-                com_openmeteo_EnsembleHourly.add(windspeed80m: offset, &fbb)
-            case .winddirection_80m:
-                com_openmeteo_EnsembleHourly.add(winddirection80m: offset, &fbb)
-            case .windspeed_120m:
-                com_openmeteo_EnsembleHourly.add(windspeed120m: offset, &fbb)
-            case .winddirection_120m:
-                com_openmeteo_EnsembleHourly.add(winddirection120m: offset, &fbb)
-            case .direct_normal_irradiance:
-                com_openmeteo_EnsembleHourly.add(directNormalIrradiance: offset, &fbb)
-            case .et0_fao_evapotranspiration:
-                com_openmeteo_EnsembleHourly.add(et0FaoEvapotranspiration: offset, &fbb)
-            case .vapor_pressure_deficit:
-                com_openmeteo_EnsembleHourly.add(vaporPressureDeficit: offset, &fbb)
-            case .shortwave_radiation:
-                com_openmeteo_EnsembleHourly.add(shortwaveRadiation: offset, &fbb)
-            case .snowfall:
-                com_openmeteo_EnsembleHourly.add(snowfall: offset, &fbb)
-            case .snow_depth:
-                com_openmeteo_EnsembleHourly.add(snowDepth: offset, &fbb)
-            case .surface_pressure:
-                com_openmeteo_EnsembleHourly.add(surfacePressure: offset, &fbb)
-            case .shortwave_radiation_instant:
-                com_openmeteo_EnsembleHourly.add(shortwaveRadiationInstant: offset, &fbb)
-            case .diffuse_radiation_instant:
-                com_openmeteo_EnsembleHourly.add(diffuseRadiationInstant: offset, &fbb)
-            case .direct_radiation_instant:
-                com_openmeteo_EnsembleHourly.add(directRadiationInstant: offset, &fbb)
-            case .direct_normal_irradiance_instant:
-                com_openmeteo_EnsembleHourly.add(directNormalIrradianceInstant: offset, &fbb)
-            case .is_day:
-                com_openmeteo_EnsembleHourly.add(isDay: offset, &fbb)
-            case .visibility:
-                com_openmeteo_EnsembleHourly.add(visibility: offset, &fbb)
-            case .freezinglevel_height:
-                com_openmeteo_EnsembleHourly.add(freezinglevelHeight: offset, &fbb)
-            case .uv_index:
-                com_openmeteo_EnsembleHourly.add(uvIndex: offset, &fbb)
-            case .uv_index_clear_sky:
-                com_openmeteo_EnsembleHourly.add(uvIndexClearSky: offset, &fbb)
-            case .cape:
-                com_openmeteo_EnsembleHourly.add(cape: offset, &fbb)
-            case .surface_temperature:
-                com_openmeteo_EnsembleHourly.add(surfaceTemperature: offset, &fbb)
-            case .soil_temperature_0_to_10cm:
-                com_openmeteo_EnsembleHourly.add(soilTemperature0To10cm: offset, &fbb)
-            case .soil_temperature_10_to_40cm:
-                com_openmeteo_EnsembleHourly.add(soilTemperature10To40cm: offset, &fbb)
-            case .soil_temperature_40_to_100cm:
-                com_openmeteo_EnsembleHourly.add(soilTemperature40To100cm: offset, &fbb)
-            case .soil_temperature_100_to_200cm:
-                com_openmeteo_EnsembleHourly.add(soilTemperature100To200cm: offset, &fbb)
-            case .soil_moisture_0_to_10cm:
-                com_openmeteo_EnsembleHourly.add(soilMoisture0To10cm: offset, &fbb)
-            case .soil_moisture_10_to_40cm:
-                com_openmeteo_EnsembleHourly.add(soilMoisture10To40cm: offset, &fbb)
-            case .soil_moisture_40_to_100cm:
-                com_openmeteo_EnsembleHourly.add(soilMoisture40To100cm: offset, &fbb)
-            case .soil_moisture_100_to_200cm:
-                com_openmeteo_EnsembleHourly.add(soilTemperature100To200cm: offset, &fbb)
-            }
-        }
-        for (pressure, offset) in offsets.pressure {
-            switch pressure {
-            case .temperature:
-                com_openmeteo_EnsembleHourly.add(pressureLevelTemperature: offset, &fbb)
-            case .geopotential_height:
-                com_openmeteo_EnsembleHourly.add(pressureLevelGeopotentialHeight: offset, &fbb)
-            case .relativehumidity:
-                com_openmeteo_EnsembleHourly.add(pressureLevelRelativehumidity: offset, &fbb)
-            case .windspeed:
-                com_openmeteo_EnsembleHourly.add(pressureLevelWindspeed: offset, &fbb)
-            case .winddirection:
-                com_openmeteo_EnsembleHourly.add(pressureLevelWinddirection: offset, &fbb)
-            case .dewpoint:
-                com_openmeteo_EnsembleHourly.add(pressureLevelDewpoint: offset, &fbb)
-            case .cloudcover:
-                com_openmeteo_EnsembleHourly.add(pressureLevelCloudcover: offset, &fbb)
-            case .vertical_velocity:
-                com_openmeteo_EnsembleHourly.add(pressureLevelVerticalVelocity: offset, &fbb)
-            }
-        }
-        return com_openmeteo_EnsembleHourly.endEnsembleHourly(&fbb, start: start)
-    }
-    
-    static func writeToFlatbuffer(section: ForecastapiResult<Self>.PerModel, _ fbb: inout FlatBufferBuilder, timezone: TimezoneWithOffset, fixedGenerationTime: Double?) throws {
-        let generationTimeStart = Date()
-        let hourly = (try section.hourly?()).map { encodeHourly(section: $0, &fbb) } ?? Offset()
-        let generationTimeMs = fixedGenerationTime ?? (Date().timeIntervalSince(generationTimeStart) * 1000)
-        
-        let result = com_openmeteo_EnsembleApi.createEnsembleApi(
-            &fbb,
-            latitude: section.latitude,
-            longitude: section.longitude,
-            elevation: section.elevation ?? .nan,
-            model: section.model.flatBufferModel,
-            generationtimeMs: Float32(generationTimeMs),
-            utcOffsetSeconds: Int32(timezone.utcOffsetSeconds),
-            timezoneOffset: fbb.create(string: timezone.identifier),
-            timezoneAbbreviationOffset: fbb.create(string: timezone.abbreviation),
-            hourlyOffset: hourly
-        )
-        fbb.finish(offset: result, addPrefix: true)
     }
 }
